Fix some mistakes
