+++ app/src/main/java/com/gks/absolutebank/MainActivity.kt
@@ -5,14 +5,13 @@ import androidx.activity.ComponentActivity
 import androidx.activity.compose.setContent
 import androidx.activity.enableEdgeToEdge
 import androidx.compose.foundation.layout.fillMaxSize
-import androidx.compose.foundation.layout.padding
 import androidx.compose.material3.Scaffold
 import androidx.compose.material3.Text
 import androidx.compose.runtime.Composable
 import androidx.compose.ui.Modifier
 import androidx.compose.ui.tooling.preview.Preview
-import com.gks.absolutebank.feature.main.MainScreenLayout
-import com.gks.absolutebank.feature.main.MainScreenViewModel
+import com.gks.absolutebank.feature.main1.MainScreenLayout
+import com.gks.absolutebank.feature.main.ui.MainScreenViewModel
 import com.gks.absolutebank.ui.theme.AbsoluteBankTheme
 
 class MainActivity : ComponentActivity() {
