+++ app/src/main/java/com/gks/absolutebank/feature/detail/ui/component/CardLayout.kt
@@ -1,0 +1,2 @@
+package com.gks.absolutebank.feature.detail.ui.component
+
