--- app/src/main/java/com/gks/absolutebank/feature/main-screen/domain/Account.kt
@@ -1,14 +1,0 @@
-package com.gks.absolutebank.feature.main
-
-import androidx.compose.runtime.Immutable
-
-@Immutable
-data class Account(
-  val id: String,
-  val number: String,
-  val balance: Double,
-  val currency: String,
-  val status: String,
-  val cards: List<Card>
-)
-
