+++ app/src/main/java/com/gks/absolutebank/feature/main/domain/entity/Card.kt
@@ -1,10 +1,11 @@
-package com.gks.absolutebank.feature.main
+package com.gks.absolutebank.feature.main.domain.entity
 
 import androidx.compose.runtime.Immutable
+import com.gks.absolutebank.feature.main1.Id
 
 @Immutable
 data class Card(
-  val id: String,
+  val id: Id,
   val number: String,
   val name: String,
   val cardType: String,
