+++ app/src/main/java/com/gks/absolutebank/feature/main/domain/entity/Deposit.kt
@@ -1,10 +1,11 @@
-package com.gks.absolutebank.feature.main
+package com.gks.absolutebank.feature.main.domain.entity
 
 import androidx.compose.runtime.Immutable
+import com.gks.absolutebank.feature.main1.Id
 
 @Immutable
 data class Deposit(
-  val id: String,
+  val id: Id,
   val name: String,
   val balance: Double,
   val currency: String,
