+++ app/src/main/java/com/gks/absolutebank/feature/main/domain/entity/Id.kt
@@ -1,0 +1,4 @@
+package com.gks.absolutebank.feature.main1
+
+@JvmInline
+value class Id (private val id: String)
