+++ app/src/main/java/com/gks/absolutebank/feature/main/ui/component/DepositLayout.kt
@@ -1,4 +1,4 @@
-package com.gks.absolutebank.feature.main
+package com.gks.absolutebank.feature.main1
 
 import androidx.compose.foundation.background
 import androidx.compose.foundation.layout.Arrangement
@@ -7,10 +7,8 @@ import androidx.compose.foundation.layout.Column
 import androidx.compose.foundation.layout.Row
 import androidx.compose.foundation.layout.Spacer
 import androidx.compose.foundation.layout.fillMaxWidth
-import androidx.compose.foundation.layout.height
 import androidx.compose.foundation.layout.padding
 import androidx.compose.foundation.layout.size
-import androidx.compose.foundation.layout.width
 import androidx.compose.foundation.shape.RoundedCornerShape
 import androidx.compose.material3.Icon
 import androidx.compose.material3.Text
@@ -21,91 +19,74 @@ import androidx.compose.ui.graphics.Color
 import androidx.compose.ui.res.painterResource
 import androidx.compose.ui.res.stringResource
 import androidx.compose.ui.text.font.FontWeight
-import androidx.compose.ui.text.style.TextAlign
 import androidx.compose.ui.unit.dp
 import androidx.compose.ui.unit.sp
 import com.gks.absolutebank.R
 
 @Composable
-fun CardLayout(
-  number: String,
-  status: String,
-  paymentSystem: String
+internal fun DepositLayout(
+  modifier: Modifier = Modifier,
+  currency: String,
+  name: String,
+  balance: Double,
+  iconResource: Int,
+  term: String,
+  rate: Double,
 ) {
   Row(
     modifier = Modifier
-      .padding(16.dp)
-      .fillMaxWidth(),
+      .fillMaxWidth()
+      .padding(16.dp),
     horizontalArrangement = Arrangement.spacedBy(16.dp)
   ) {
     Box(
       modifier = Modifier
         .background(
-          color = Color.Transparent,
+          color = Color(0xFF403A47),
           shape = RoundedCornerShape(999.dp)
         )
         .size(40.dp),
       contentAlignment = Alignment.Center
     ) {
       Icon(
-        painter = painterResource(R.drawable.ic_input_24),
+        painter = painterResource(iconResource),
         tint = Color.White,
         contentDescription = null
       )
     }
     Column(
-      verticalArrangement = Arrangement.spacedBy(2.dp)
+      verticalArrangement = Arrangement.spacedBy(4.dp)
     ) {
       Text(
-        text = number,
+        text = name,
         color = Color.White,
         fontSize = 15.sp,
         fontWeight = FontWeight(400)
       )
       Text(
-        text = status,
-        color = Color.White,
-        fontSize = 13.sp,
+        text = balance.toString() + " " + currency,
+        color = Color(0xFF6C78E6),
+        fontSize = 15.sp,
         fontWeight = FontWeight(400)
       )
     }
     Spacer(modifier = Modifier.weight(1f))
-    CardView(
-      number = number.takeLast(4),
-      iconResource = if (paymentSystem == stringResource(R.string.visa))
-        R.drawable.ic_visa_16_6
-      else R.drawable.ic_mastercard_16_12
+    Column(
+      verticalArrangement = Arrangement.spacedBy(7.dp)
+    ) {
+      Text(
+        text = stringResource(R.string.rate) + " " + rate + "%",
+        color = Color(0xFF706D76),
+        fontSize = 11.sp,
+        fontWeight = FontWeight(400)
       )
+      Text(
+        text = stringResource(R.string.up_to) + term,
+        color = Color(0xFF706D76),
+        fontSize = 11.sp,
+        fontWeight = FontWeight(400)
+      )
+    }
   }
 }
 
-@Composable
-fun CardView(
-  number: String,
-  iconResource: Int
-) {
-  Box(
-    modifier = Modifier
-      .background(
-        color = Color.Gray,
-        shape = RoundedCornerShape(2.dp)
-      )
-      .width(40.dp)
-      .height(28.dp)
-  ) {
-    Text(
-      modifier = Modifier
-        .fillMaxWidth()
-        .padding(horizontal = 2.dp),
-      text = number,
-      fontSize = 10.sp,
-      color = Color.White,
-      textAlign = TextAlign.Right
-    )
-    Icon(
-      painter = painterResource(iconResource),
-      tint = Color.Transparent,
-      contentDescription = null
-    )
-  }
-}
