+++ app/src/main/java/com/gks/absolutebank/ui/theme/Type.kt
@@ -10,12 +10,11 @@ import androidx.compose.ui.unit.sp
 val Typography = Typography(
   bodyLarge = TextStyle(
     fontFamily = FontFamily.Default,
-    fontWeight = FontWeight.Normal,
     fontSize = 16.sp,
+    fontWeight = FontWeight(600),
     lineHeight = 24.sp,
     letterSpacing = 0.5.sp
-  )
-  /* Other default text styles to override
+  ),
     titleLarge = TextStyle(
         fontFamily = FontFamily.Default,
         fontWeight = FontWeight.Normal,
@@ -29,6 +28,7 @@ val Typography = Typography(
         fontSize = 11.sp,
         lineHeight = 16.sp,
         letterSpacing = 0.5.sp
-    )
-    */
+    ),
+
+
 )
